Make readme good enough to merge
Sorry for doing PRs inside of PRs but I added some of what we talked about in ynoproject/ynotranslations/pull/181 since you said you'll be busy and I figured it would be easier to just get everything down and then worry about other issues later, since now you can just post this in the Discord and have people give feedback on stuff like the credits and also allow other people to potentionally improve this in the meantime. Sorry if it feels like this is rushing you though, feel free to ignore this until you have time to work on this if it's too much to ask for now.